Move `PatternRelationship` to IR module
Also move `PatternLength` trait and implementations.
