+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/ast/convert/plannerQuery/PatternConverters.scala
@@ -22,9 +22,9 @@ package org.neo4j.cypher.internal.compiler.v3_2.ast.convert.plannerQuery
 import org.neo4j.cypher.internal.compiler.v3_2.ast.convert.plannerQuery.ExpressionConverters._
 import org.neo4j.cypher.internal.compiler.v3_2.helpers.FreshIdNameGenerator
 import org.neo4j.cypher.internal.compiler.v3_2.planner.CantHandleQueryException
-import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans.{PatternRelationship, ShortestPathPattern}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans.ShortestPathPattern
 import org.neo4j.cypher.internal.frontend.v3_2.ast._
-import org.neo4j.cypher.internal.ir.v3_2.IdName
+import org.neo4j.cypher.internal.ir.v3_2.{IdName, PatternRelationship}
 
 object PatternConverters {
