+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/planner/logical/idp/joinSolverStep.scala
@@ -21,9 +21,9 @@ package org.neo4j.cypher.internal.compiler.v3_2.planner.logical.idp
 
 import org.neo4j.cypher.internal.compiler.v3_2.planner.QueryGraph
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.idp.joinSolverStep._
-import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans.{LogicalPlan, PatternRelationship}
+import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.plans.LogicalPlan
 import org.neo4j.cypher.internal.compiler.v3_2.planner.logical.{LogicalPlanningContext, LogicalPlanningSupport}
-import org.neo4j.cypher.internal.ir.v3_2.IdName
+import org.neo4j.cypher.internal.ir.v3_2.{IdName, PatternRelationship}
 
 object joinSolverStep {
   val VERBOSE = false
